pyproject.toml
[project]
name = "hiyobot"
version = "6.0.0"
description = ""
authors = [
    {name = "Ryu Juheon",email = "saidbysolo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "discord-py (>=2.6.3,<3.0.0)",
    "delphinium (>=1.4.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
